Add A* search algorithm option for path solving
The shortest_path now has the option to solve with either
Dijkstra or A*. A* offers some performance advantage since
is relies on the distance heuristic. Existing code is
unchanged since the algorithm choice is an optional parameter,
defaulting to Dijkstra. Example 2 script is modified to show
how it is used and a speed comparison.

Output from examples/2_calculate_shortest_distance.py:

   Solved with Dijkstra's algorithm in 2.5208137035369873 seconds.
   Solved with A* algorithm in 1.7424323558807373 seconds.
   Shortest path distance: 16215.660241221995